dbmsBackend/requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.2
Django==3.2.22
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
importlib-metadata==4.8.3
kombu==4.6.11
Markdown==3.3.7
psycopg2-binary==2.9.8
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.1.1
vine==1.3.0
zipp==3.6.0
